# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/sound": {
            "post": {
                "tags": [
                    "Sounds"
                ],
                "summary": "Create a sound to insert into DB",
                "operationId": "2156b690b5da05ff3098abd759136baa",
                "requestBody": {
                    "description": " the description should be at least 4 character.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Sound"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/sound/{id}": {
            "get": {
                "tags": [
                    "Sounds"
                ],
                "summary": "get one resource",
                "operationId": "d6e6fa87e3f92ef584fba6e8231cdfc2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sound's id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sound"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sounds"
                ],
                "summary": "edit a resource from db",
                "operationId": "3a53c40f8dbf54329dbcf3dbda7e7de0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sounds id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": " the description should be at least 4 character.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Sound"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sound"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sounds"
                ],
                "summary": "edit a resource from db",
                "operationId": "5541ca8ad98054311d3ff618e2201dea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sounds id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Sound": {
                "title": "Sound",
                "description": "Sound model",
                "properties": {
                    "description": {
                        "title": "description",
                        "description": "description of the Sound",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Sound"
                }
            },
...
        }
    }
}